+++ src/WebCompiler/Compile/SassCompiler.cs
@@ -1,10 +1,24 @@
-﻿using System;
+﻿using Newtonsoft.Json.Linq;
+using System;
+using System.Diagnostics;
 using System.IO;
+using System.Text;
+using System.Text.RegularExpressions;
 
 namespace WebCompiler
 {
-    internal class SassCompiler : ICompiler
+    class SassCompiler : ICompiler
     {
+        private static Regex _errorRx = new Regex("(?<message>.+) on line (?<line>[0-9]+), column (?<column>[0-9]+)", RegexOptions.Compiled);
+        private string _path;
+        private string _output = string.Empty;
+        private string _error = string.Empty;
+
+        public SassCompiler(string path)
+        {
+            _path = path;
+        }
+
         public CompilerResult Compile(Config config)
         {
             string baseFolder = Path.GetDirectoryName(config.FileName);
@@ -13,46 +27,93 @@ namespace WebCompiler
             FileInfo info = new FileInfo(inputFile);
             string content = File.ReadAllText(info.FullName);
 
-            string sourceMap = config.SourceMap ? config.GetAbsoluteOutputFile() + ".map" : null;
-
             CompilerResult result = new CompilerResult
             {
                 FileName = info.FullName,
                 OriginalContent = content,
             };
 
-            SassOptions options = new SassOptions(config);
-
             try
             {
-                LibSassNet.SassCompiler compiler = new LibSassNet.SassCompiler();
-                var compilerResult = compiler.CompileFile(inputFile, (LibSassNet.OutputStyle)options.OutputStyle, sourceMap, options.IncludeSourceComments, options.Precision);
-                result.CompiledContent = compilerResult.CSS;
-                result.SourceMap = compilerResult.SourceMap;
+                RunCompilerProcess(config, info);
+
+                result.CompiledContent = _output;
+
+                if (_error.Length > 0)
+                {
+                    JObject json = JObject.Parse(_error);
+
+                    CompilerError ce = new CompilerError
+                    {
+                        FileName = info.FullName,
+                        Message = json["message"].ToString(),
+                        ColumnNumber = int.Parse(json["column"].ToString()),
+                        LineNumber = int.Parse(json["line"].ToString()),
+                    };
+
+                    result.Errors.Add(ce);
+                }
             }
             catch (Exception ex)
             {
                 CompilerError error = new CompilerError
                 {
                     FileName = info.FullName,
-                    Message = ex.Message.Replace("/", "\\").Replace(info.FullName, string.Empty).Trim()
+                    Message = string.IsNullOrEmpty(_error) ? ex.Message : _error,
+                    LineNumber = 0,
+                    ColumnNumber = 0,
                 };
-
-                if (error.Message.StartsWith(":"))
-                {
-                    int end = error.Message.IndexOf(':', 1);
-                    int line = 0;
-                    if (int.TryParse(error.Message.Substring(1, end - 1), out line))
-                    {
-                        error.LineNumber = line;
-                        error.Message = error.Message.Substring(end + 1).Trim();
-                    }
-                }
 
                 result.Errors.Add(error);
             }
 
             return result;
+        }
+
+        private void RunCompilerProcess(Config config, FileInfo info)
+        {
+            string arguments = ConstructArguments(config);
+
+            ProcessStartInfo start = new ProcessStartInfo
+            {
+                WorkingDirectory = info.Directory.FullName,
+                UseShellExecute = false,
+                WindowStyle = ProcessWindowStyle.Hidden,
+                CreateNoWindow = true,
+                FileName = "cmd.exe",
+                Arguments = $"/c \"\"{Path.Combine(_path, "node_modules\\.bin\\node-sass.cmd")}\" {arguments} \"{info.FullName}\"\"",
+                StandardOutputEncoding = Encoding.UTF8,
+                StandardErrorEncoding = Encoding.UTF8,
+                RedirectStandardOutput = true,
+                RedirectStandardError = true,
+            };
+
+            start.EnvironmentVariables["PATH"] = _path + ";" + start.EnvironmentVariables["PATH"];
+
+            Process p = Process.Start(start);
+            var stdout = p.StandardOutput.ReadToEndAsync();
+            var stderr = p.StandardError.ReadToEndAsync();
+            p.WaitForExit();
+
+            _output = stdout.Result;
+            _error = stderr.Result;
+        }
+
+        private static string ConstructArguments(Config config)
+        {
+            string arguments = " --no-color";
+
+            if (config.SourceMap)
+                arguments += " --source-map-embed";
+
+            SassOptions options = new SassOptions(config);
+
+            arguments += " --precision=" + options.Precision;
+
+            if (!string.IsNullOrEmpty(options.OutputStyle))
+                arguments += " --output-style=" + options.OutputStyle;
+
+            return arguments;
         }
     }
 }