+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [x] Task Runner Explorer integration
 - [x] Command line support
 - [x] Optimize compilation of newly added configs (#72)
+- [x] Use `node-sass` to compile Sass files (#74)
 - [ ] Generate gulpfile.js from compilerconfig.json (#34)
 - [ ] Preview window (#6)
 - [ ] File globbing pattern support (#49)